[Tabs] Let each tab set its own width
The Tab component has a new prop of type number: labelWidth. This number
is the width of the label in pixels. The user must either set this prop
to every tab or none. If this condition is not met, the ink bar is not
drawn, and a warning is displayed in the console.

The Tabs directory has a new file: TabUtils.js. For now this file
has a single function: calculateTabWidthAndInkbarPosition which is
necessary now that widths are variable. A test for this function has
been written in the TabUtils.spec.js file.

The docs have also been updated with information and an example for this
prop.

This solves issue #4420. This approach was suggested in #5301.